perf: rename utils passwd to bcrypt
